Moved the logic of the coins data to the web worker
