+++ src/redux/actions/coinsApiActions.js
@@ -2,7 +2,7 @@ import { CALL_API } from 'redux-api-middleware';
 import * as Actions from '../ActionNames';
 import config from 'config';
 
-export const fetchCoinsData = (locale) => ({
+export const fetchCoinsData = (locale, targetCurrency) => ({
   [CALL_API]: {
     endpoint: `${config.ORIGINS.COINS_IO}/front`,
     method: 'GET',
@@ -10,7 +10,11 @@ export const fetchCoinsData = (locale) => ({
       Actions.FETCH_COINS_DATA,
       {
         type: Actions.FETCH_COINS_DATA_SUCCESS,
-        meta: { locale }
+        meta: {
+          locale,
+          targetCurrency,
+          WebWorker: true
+        }
       },
       Actions.FETCH_COINS_DATA_FAILURE
     ]
