+++ src/redux/middlewares/worker.js
@@ -1,4 +1,4 @@
 import createWorkerMiddleware from './createWorkerMiddleware';
-const CoinsWorker = require('worker-loader?inline!../workers/CoinsWorker');
+const CoinsWorker = require('worker-loader?!../workers/CoinsWorker');
 
 export default createWorkerMiddleware(new CoinsWorker());
